refactor: rename to snake case
