+++ projects/portfolio/src/client/components/chat/chat-main.tsx
@@ -1,14 +1,14 @@
 import { ChatInput } from '#/client/components/chat/chat-input'
 import { PromptTemplate, type TemplateInput } from '#/client/components/chat/prompt-template'
 import { FileImageInput, FileImagePreview } from '#/client/components/input/file-image-input'
-import { ArrowUpIcon } from '#/client/components/svg/ArrowUpIcon'
-import { ChatbotIcon } from '#/client/components/svg/ChatbotIcon'
-import { CheckIcon } from '#/client/components/svg/CheckIcon'
-import { CopyIcon } from '#/client/components/svg/CopyIcon'
-import { DeleteIcon } from '#/client/components/svg/DeleteIcon'
-import { SpinnerIcon } from '#/client/components/svg/SpinnerIcon'
-import { StopIcon } from '#/client/components/svg/StopIcon'
-import { UploadIcon } from '#/client/components/svg/UploadIcon'
+import { ArrowUpIcon } from '#/client/components/svg/arrow-upIcon'
+import { ChatbotIcon } from '#/client/components/svg/chatbot-icon'
+import { CheckIcon } from '#/client/components/svg/check-icon'
+import { CopyIcon } from '#/client/components/svg/copy-icon'
+import { DeleteIcon } from '#/client/components/svg/delete-icon'
+import { SpinnerIcon } from '#/client/components/svg/spinner-icon'
+import { StopIcon } from '#/client/components/svg/stop-icon'
+import { UploadIcon } from '#/client/components/svg/upload-icon'
 import type { Settings } from '#/client/lib/remoteStorageSettings'
 import type { AppType } from '#/server/app.d'
 import type { Conversation } from '#/types'
