+++ projects/portfolio/src/client/components/svg/about-icon.tsx
