+++ projects/portfolio/src/client/components/svg/chatbot-icon.tsx
