+++ projects/portfolio/src/client/components/svg/close-icon.tsx
