+++ projects/portfolio/src/client/components/svg/copy-icon.tsx
