+++ projects/portfolio/src/client/components/svg/dashboard-icon.tsx
