+++ projects/portfolio/src/client/components/svg/delete-icon.tsx
