+++ projects/portfolio/src/client/components/svg/gear-icon.tsx
