+++ projects/portfolio/src/client/components/svg/llm-icon.tsx
