+++ projects/portfolio/src/client/components/svg/new-chat-icon.tsx
