+++ projects/portfolio/src/client/components/svg/spinner-icon.tsx
