+++ projects/portfolio/src/client/components/svg/stop-icon.tsx
