+++ projects/portfolio/src/client/components/svg/upload-icon.tsx
