+++ projects/portfolio/src/client/routes/__root.tsx
@@ -1,8 +1,8 @@
 import { AppLayout } from '#/client/components/app-layout'
-import { AboutIcon } from '#/client/components/svg/AboutIcon'
-import { ChatbotIcon } from '#/client/components/svg/ChatbotIcon'
-import { DashboardIcon } from '#/client/components/svg/DashboardIcon'
-import { LlmIcon } from '#/client/components/svg/LlmIcon'
+import { AboutIcon } from '#/client/components/svg/about-icon'
+import { ChatbotIcon } from '#/client/components/svg/chatbot-icon'
+import { DashboardIcon } from '#/client/components/svg/dashboard-icon'
+import { LlmIcon } from '#/client/components/svg/llm-icon'
 import { createRootRoute, Outlet } from '@tanstack/react-router'
 import React from 'react'
